+++ src/apps/viewer/extensions.ts
@@ -30,6 +30,10 @@ import { PluginSpec } from '../../mol-plugin/spec';
 import { MVSData } from '../../extensions/mvs/mvs-data';
 import * as MVSUtil from '../../extensions/mvs/util';
 import { KinemageExtension } from '../../extensions/kinemage/behavior';
+import * as interactivity from '../../extensions/plugin/interactivity';
+import * as loaders from '../../extensions/plugin/load';
+import { PluginViewModel } from '../../extensions/plugin/view-model';
+import { PluginUIViewModel } from '../../extensions/plugin/ui-view-model';
 
 export const ExtensionMap = {
     // Mol* built-in extensions
@@ -70,5 +74,13 @@ export const PluginExtensions = {
         qualityAssessment: {
             config: MAQualityAssessmentConfig
         }
-    }
+    },
+    plugin: {
+        interactivity,
+        loaders,
+        models: {
+            PluginViewModel,
+            PluginUIViewModel,
+        },
+    },
 };
